Extracted hsv and rgb transforms
Working on a non-Arduino, bare metal build that generally works for rp2040
and STM32 devices (Cortex M0).

Unfortunately as I started adding tests, it became clear that the
conversion functions included in this library are not even close to
desired color values.

For example, CYAN becomes 100% blue. Magenta becomes 100% red.